Move extract tests to a separate file
