+++ tests/tests.py
@@ -4,7 +4,6 @@ import cohere
 import string
 import random
 from cohere.classify import Example
-from cohere.extract import Entity, Example as ExtractExample
 
 API_KEY = os.getenv('CO_API_KEY')
 assert type(API_KEY)
@@ -238,118 +237,6 @@ class TestClassify(unittest.TestCase):
         ], 'this is a classifier to determine if a word is a fruit of a color', 'This is a')
         self.assertEqual(prediction.classifications[0].prediction, 'fruit')
         self.assertEqual(prediction.classifications[1].prediction, 'color')
-
-
-class TestExtract(unittest.TestCase):
-    def test_success(self):
-        examples = [ExtractExample(
-            text="hello my name is John, and I like to play ping pong",
-            entities=[Entity(type="Name", value="John")])]
-        texts = ["hello Roberta, how are you doing today?"]
-
-        extractions = co.extract('small', examples, texts)
-
-        self.assertIsInstance(extractions, list)
-        self.assertIsInstance(extractions[0].text, str)
-        self.assertIsInstance(extractions[0].entities, list)
-        self.assertEqual(extractions[0].entities[0].type, "Name")
-        self.assertEqual(extractions[0].entities[0].value, "Roberta")
-
-    def test_empty_text(self):
-        with self.assertRaises(cohere.CohereError):
-            co.extract(
-                'small', examples=[ExtractExample(
-                    text="hello my name is John, and I like to play ping pong",
-                    entities=[Entity(type="Name", value="John")])],
-                texts=[""])
-
-    def test_empty_entities(self):
-        with self.assertRaises(cohere.CohereError):
-            co.extract(
-                'large', examples=[ExtractExample(
-                    text="hello my name is John, and I like to play ping pong",
-                    entities=[])],
-                texts=["hello Roberta, how are you doing today?"])
-
-    def test_varying_amount_of_entities(self):
-        examples = [
-            ExtractExample(
-                text="the bananas are red",
-                entities=[Entity(type="fruit", value="bananas"), Entity(type="color", value="red")]),
-            ExtractExample(
-                text="i love the color blue",
-                entities=[Entity(type="color", value="blue")]),
-            ExtractExample(
-                text="i love apples",
-                entities=[Entity(type="fruit", value="apple")]),
-            ExtractExample(
-                text="purple is my favorite color",
-                entities=[Entity(type="color", value="purple")]),
-            ExtractExample(
-                text="wow, that apple is green?",
-                entities=[Entity(type="fruit", value="apple"), Entity(type="color", value="green")])]
-        texts = ["i love bananas", "my favorite color is yellow", "i love green apples"]
-
-        extractions = co.extract('medium', examples, texts)
-
-        self.assertIsInstance(extractions, list)
-        self.assertIsInstance(extractions[0].text, str)
-        self.assertIsInstance(extractions[1].text, str)
-        self.assertIsInstance(extractions[2].text, str)
-        self.assertIsInstance(extractions[0].entities, list)
-        self.assertIsInstance(extractions[1].entities, list)
-        self.assertIsInstance(extractions[2].entities, list)
-        self.assertEqual(len(extractions[0].entities), 1)
-        self.assertEqual(len(extractions[1].entities), 1)
-        self.assertEqual(len(extractions[2].entities), 2)
-
-    def test_many_examples_and_multiple_texts(self):
-        examples = [
-            ExtractExample(
-                text="hello my name is John, and I like to play ping pong",
-                entities=[Entity(type="Name", value="John"), Entity(type="Game", value="ping pong")]),
-            ExtractExample(
-                text="greetings, I'm Roberta and I like to play golf",
-                entities=[Entity(type="Name", value="Roberta"), Entity(type="Game", value="golf")]),
-            ExtractExample(
-                text="let me introduce myself, my name is Tina and I like to play baseball",
-                entities=[Entity(type="Name", value="Tina"), Entity(type="Game", value="baseball")])]
-        texts = ["hi, my name is Charlie and I like to play basketball", "hello, I'm Olivia and I like to play soccer"]
-
-        extractions = co.extract('medium', examples, texts)
-
-        self.assertEqual(len(extractions), 2)
-        self.assertIsInstance(extractions, list)
-        self.assertIsInstance(extractions[0].text, str)
-        self.assertIsInstance(extractions[1].text, str)
-        self.assertIsInstance(extractions[0].entities, list)
-        self.assertIsInstance(extractions[1].entities, list)
-        self.assertEqual(len(extractions[0].entities), 2)
-        self.assertEqual(len(extractions[1].entities), 2)
-
-    def test_no_entities(self):
-        examples = [
-            ExtractExample(
-                text="hello my name is John, and I like to play ping pong",
-                entities=[Entity(type="Name", value="John"), Entity(type="Game", value="ping pong")]),
-            ExtractExample(
-                text="greetings, I'm Roberta and I like to play golf",
-                entities=[Entity(type="Name", value="Roberta"), Entity(type="Game", value="golf")]),
-            ExtractExample(
-                text="let me introduce myself, my name is Tina and I like to play baseball",
-                entities=[Entity(type="Name", value="Tina"), Entity(type="Game", value="baseball")])]
-        texts = ["hi, my name is Charlie and I like to play basketball", "hello!"]
-
-        extractions = co.extract('medium', examples, texts)
-
-        self.assertEqual(len(extractions), 2)
-        self.assertIsInstance(extractions, list)
-        self.assertIsInstance(extractions[0].text, str)
-        self.assertIsInstance(extractions[1].text, str)
-        self.assertIsInstance(extractions[0].entities, list)
-        self.assertIsInstance(extractions[1].entities, list)
-        self.assertEqual(len(extractions[0].entities), 2)
-        self.assertEqual(len(extractions[1].entities), 0)
 
 
 class TestTokenize(unittest.TestCase):
