+++ tests/utils.py
@@ -1,0 +1,7 @@
+import os
+
+
+def get_api_key() -> str:
+    api_key = os.getenv('CO_API_KEY')
+    assert api_key, 'CO_API_KEY environment variable not set'
+    return api_key
